waba-backend/requirements.txt
africastalking==1.2.9
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
influxdb==5.3.2
influxdb-client==1.49.0
msgpack==1.1.0
packaging==25.0
paho-mqtt==2.1.0
pillow==11.2.1
psycopg2==2.9.10
pyasn1==0.4.8
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
reactivex==4.0.4
requests==2.32.3
rsa==4.9.1
schema==0.7.7
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.4.0
